split user group mocks into separate files
