+++ src/mocks/handlers/user-group/detail.handlers.ts
@@ -1,30 +1,25 @@
 const { rest } = window.MockServiceWorker;
-import { umbUserGroupData } from '../data/user-group.data.js';
+import { umbUserGroupData } from '../../data/user-group.data.js';
+import { slug } from './slug.js';
 import { umbracoPath } from '@umbraco-cms/backoffice/utils';
 
-const slug = '/user-group';
+export const detailHandlers = [
+	rest.post(umbracoPath(`${slug}`), async (req, res, ctx) => {
+		const data = await req.json();
+		if (!data) return;
 
-export const handlers = [
-	rest.get(umbracoPath(`${slug}/item`), (req, res, ctx) => {
-		const ids = req.url.searchParams.getAll('id');
-		if (!ids) return;
-		const items = umbUserGroupData.getItems(ids);
+		umbUserGroupData.insert(data);
 
-		return res(ctx.status(200), ctx.json(items));
-	}),
-
-	rest.get(umbracoPath(`${slug}`), (req, res, ctx) => {
-		const response = umbUserGroupData.getAll();
-
-		return res(ctx.status(200), ctx.json(response));
+		return res(ctx.status(200));
 	}),
 
 	rest.get(umbracoPath(`${slug}/:id`), (req, res, ctx) => {
 		const id = req.params.id as string;
 		if (!id) return;
-		const userGroup = umbUserGroupData.getById(id);
 
-		return res(ctx.status(200), ctx.json(userGroup));
+		const dataType = umbUserGroupData.getById(id);
+
+		return res(ctx.status(200), ctx.json(dataType));
 	}),
 
 	rest.put(umbracoPath(`${slug}/:id`), async (req, res, ctx) => {
