+++ src/mocks/handlers/user-group/slug.ts
@@ -1,0 +1,1 @@
+export const slug = '/user-group';
